Modify the reset shell script files on July 7, 2025.

+++ reset/reset_opencloudos/reset_opencloudos.sh
@@ -3,7 +3,7 @@
 #**********************************************************************************
 #Author:        Raymond
 #QQ:            88563128
-#Date:          2025-04-08
+#Date:          2025-06-10
 #FileName:      reset_opencloudos.sh
 #MIRROR:        https://wx.zsxq.com/group/15555885545422
 #Description:   The reset linux system initialization script supports 
@@ -14,26 +14,29 @@ COLOR="echo -e \\033[01;31m"
 END='\033[0m'
 
 os(){
-    OS_ID=`sed -rn '/^NAME=/s@.*="([[:alpha:]]+).*"$@\1@p' /etc/os-release`
-    OS_RELEASE=`sed -rn '/^VERSION_ID=/s@.*="?([0-9.]+)"?@\1@p' /etc/os-release`
-    OS_RELEASE_VERSION=`sed -rn '/^VERSION_ID=/s@.*="?([0-9]+)\.?.*"?@\1@p' /etc/os-release`
+    . /etc/os-release
+    MAIN_NAME=`sed -rn '/^NAME=/s@.*="([[:alpha:]]+).*"$@\1@p' /etc/os-release`
+    MAIN_VERSION_ID=`sed -rn '/^VERSION_ID=/s@.*="?([0-9]+)\.?.*"?@\1@p' /etc/os-release`
+    if [ ${MAIN_NAME} == "Ubuntu" -o ${MAIN_NAME} == "Debian" ];then
+        FULL_NAME="${PRETTY_NAME}"
+    elif [ ${MAIN_NAME} == "UOS" ];then
+        FULL_NAME="${NAME}"
+    else
+        FULL_NAME="${NAME} ${VERSION_ID}"
+    fi
 }
 
 set_opencloudos_7_8_eth(){
-    if grep -Eqi "(net\.ifnames|biosdevname)" /etc/default/grub;then
-        ${COLOR}"${OS_ID} ${OS_RELEASE} 网卡名配置文件已修改,不用修改!"${END}
+    sed -ri.bak '/^GRUB_CMDLINE_LINUX=/s@"$@ net.ifnames=0 biosdevname=0"@' /etc/default/grub
+    if lsblk | grep -q efi;then
+        EFI_DIR=`find /boot/efi/ -name "grub.cfg" | awk -F"/" '{print $5}'`
+        grub2-mkconfig -o /boot/efi/EFI/${EFI_DIR}/grub.cfg >& /dev/null
     else
-        sed -ri.bak '/^GRUB_CMDLINE_LINUX=/s@"$@ net.ifnames=0 biosdevname=0"@' /etc/default/grub
-        if lsblk | grep -q efi;then
-            EFI_DIR=`find /boot/efi/ -name "grub.cfg" | awk -F"/" '{print $5}'`
-            grub2-mkconfig -o /boot/efi/EFI/${EFI_DIR}/grub.cfg >& /dev/null
-        else
-            grub2-mkconfig -o /boot/grub2/grub.cfg >& /dev/null
-        fi
-        ETHNAME=`ip addr | awk -F"[ :]" '/^2/{print $3}'`
-        mv /etc/sysconfig/network-scripts/ifcfg-${ETHNAME} /etc/sysconfig/network-scripts/ifcfg-eth0
-        sed -i.bak 's/'${ETHNAME}'/eth0/' /etc/sysconfig/network-scripts/ifcfg-eth0
-    fi   
+        grub2-mkconfig -o /boot/grub2/grub.cfg >& /dev/null
+    fi
+    ETHNAME=`ip addr | awk -F"[ :]" '/^2/{print $3}'`
+    mv /etc/sysconfig/network-scripts/ifcfg-${ETHNAME} /etc/sysconfig/network-scripts/ifcfg-eth0
+    sed -i.bak 's/'${ETHNAME}'/eth0/' /etc/sysconfig/network-scripts/ifcfg-eth0  
 }
 
 set_opencloudos_9_eth0(){
@@ -65,19 +68,35 @@ EOF
 }
 
 set_eth(){
-    if [ ${OS_RELEASE_VERSION} == "7" -o ${OS_RELEASE_VERSION} == "8" ];then
-        set_opencloudos_7_8_eth
+    if [ ${MAIN_VERSION_ID} == "7" -o ${MAIN_VERSION_ID} == "8" ];then
+        if grep -Eqi "(net\.ifnames|biosdevname)" /etc/default/grub;then
+            ${COLOR}"${FULL_NAME}操作系统，网卡名配置文件已修改，不用修改！"${END}
+        else
+            set_opencloudos_7_8_eth
+	        ${COLOR}"${FULL_NAME}操作系统，网卡名已修改成功，10秒后，机器会自动重启！"${END}
+            sleep 10 && shutdown -r now
+        fi
     else
         IP_NUM=`ip addr | awk -F"[: ]" '{print $1}' | grep -v '^$' | wc -l`
         if [ ${IP_NUM} == "2" ];then
-            set_opencloudos_9_eth0
+            if [ -f /etc/systemd/network/70-eth0.link ];then
+                ${COLOR}"${FULL_NAME}操作系统，网卡名配置文件已修改，不用修改！"${END}
+            else
+                set_opencloudos_9_eth0
+	            ${COLOR}"${FULL_NAME}操作系统，网卡名已修改成功，10秒后，机器会自动重启！"${END}
+                sleep 10 && shutdown -r now
+            fi
         else
-            set_opencloudos_9_eth0
-            set_opencloudos_9_eth1
+            if [ -f /etc/systemd/network/70-eth0.link -a -f /etc/systemd/network/70-eth1.link ];then
+                ${COLOR}"${FULL_NAME}操作系统，网卡名配置文件已修改，不用修改！"${END}
+            else
+                set_opencloudos_9_eth0
+                set_opencloudos_9_eth1
+	            ${COLOR}"${FULL_NAME}操作系统，网卡名已修改成功，10秒后，机器会自动重启！"${END}
+                sleep 10 && shutdown -r now
+            fi
         fi
     fi
-	${COLOR}"${OS_ID} ${OS_RELEASE} 网卡名已修改成功,10秒后,机器会自动重启!"${END}
-    sleep 10 && shutdown -r now
 }
 
 check_ip(){
@@ -120,7 +139,7 @@ set_network_eth0(){
         check_ip ${BACKUP_DNS}
         [ $? -eq 0 ] && break
     done
-    if [ ${OS_RELEASE_VERSION} == "7" -o ${OS_RELEASE_VERSION} == "8" ];then
+    if [ ${MAIN_VERSION_ID} == "7" -o ${MAIN_VERSION_ID} == "8" ];then
         cat > /etc/sysconfig/network-scripts/ifcfg-eth0 <<-EOF
 NAME=${ETHNAME}
 DEVICE=${ETHNAME}
@@ -156,7 +175,7 @@ set_network_eth1(){
         [ $? -eq 0 ] && break
     done
     read -p "请输入子网掩码位数: " PREFIX2
-    if [ ${OS_RELEASE_VERSION} == "7" -o ${OS_RELEASE_VERSION} == "8" ];then
+    if [ ${MAIN_VERSION_ID} == "7" -o ${MAIN_VERSION_ID} == "8" ];then
         cat > /etc/sysconfig/network-scripts/ifcfg-${ETHNAME2} <<-EOF
 NAME=${ETHNAME2}
 DEVICE=${ETHNAME2}
@@ -189,19 +208,19 @@ set_network(){
         set_network_eth0
         set_network_eth1
     fi
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 网络已设置成功,请重新启动系统后生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，网络已设置成功，请重新启动系统后生效！"${END}
 }
 
 set_hostname(){
     read -p "请输入主机名: " HOST
     hostnamectl set-hostname ${HOST}
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 主机名设置成功,请重新登录生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，主机名设置成功，请重新登录生效！"${END}
 }
 
 set_hostname(){
     read -p "请输入主机名: " HOST
     hostnamectl set-hostname ${HOST}
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 主机名设置成功,请重新登录生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，主机名设置成功，请重新登录生效！"${END}
 }
 
 tencent(){
@@ -228,9 +247,9 @@ set_yum(){
     else
         sed -i -e 's|'${OLD_MIRROR}'|'${MIRROR}'|g' /etc/yum.repos.d/OpenCloudOS*.repo
     fi
-    ${COLOR}"更新镜像源中,请稍等..."${END}
+    ${COLOR}"更新镜像源中，请稍等......"${END}
     dnf clean all &> /dev/null && dnf makecache &> /dev/null
-    ${COLOR}"${OS_ID} ${OS_RELEASE} YUM源设置完成!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，镜像源设置完成！"${END}
 }
 
 base_menu(){
@@ -267,46 +286,85 @@ EOF
             break
             ;;
         *)
-            ${COLOR}"输入错误,请输入正确的数字(1-5)!"${END}
+            ${COLOR}"输入错误，请输入正确的数字(1-5)！"${END}
+            ;;
+        esac
+    done
+}
+
+set_powertools_opencloudos_8(){
+    dnf config-manager --set-enabled PowerTools
+    ${COLOR}"更新镜像源中，请稍等......"${END}
+    dnf clean all &> /dev/null && dnf makecache &> /dev/null
+    ${COLOR}"${FULL_NAME}操作系统，PowerTools仓库镜像源设置完成！"${END}
+}
+
+opencloudos_menu(){
+    while true;do
+        echo -e "\E[$[RANDOM%7+31];1m"
+        cat <<-EOF
+1)base仓库
+2)启用OpenCloudOS 8 PowerTools仓库
+3)退出
+EOF
+        echo -e '\E[0m'
+
+        read -p "请输入镜像源编号(1-3): " NUM
+        case ${NUM} in
+        1)
+            base_menu
+            ;;
+        2)
+            if [ ${MAIN_VERSION_ID} == "8" ];then
+                set_powertools_opencloudos_8
+            else
+                ${COLOR}"${FULL_NAME}操作系统，没有PowerTools仓库，不用设置！"${END}
+            fi
+            ;;
+        3)
+            break
+            ;;
+        *)
+            ${COLOR}"输入错误，请输入正确的数字(1-3)！"${END}
             ;;
         esac
     done
 }
 
 minimal_install(){
-    ${COLOR}'开始安装“Minimal安装建议安装软件包”,请稍等......'${END}
+    ${COLOR}'开始安装“Minimal安装建议安装软件包”，请稍等......'${END}
     yum -y install gcc make autoconf gcc-c++ glibc glibc-devel pcre pcre-devel openssl openssl-devel systemd-devel zlib-devel vim lrzsz tree tmux lsof tcpdump wget net-tools iotop bc bzip2 zip unzip nfs-utils man-pages &> /dev/null
-    ${COLOR}"${OS_ID} ${OS_RELEASE} Minimal安装建议安装软件包已安装完成!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，Minimal安装建议安装软件包已安装完成！"${END}
 }
 
 disable_firewalls(){
-    rpm -q firewalld &> /dev/null && { systemctl disable --now firewalld &> /dev/null; ${COLOR}"${OS_ID} ${OS_RELEASE} Firewall防火墙已关闭!"${END}; } || ${COLOR}"${OS_ID} ${OS_RELEASE} iptables防火墙已关闭!"${END}
+    rpm -q firewalld &> /dev/null && { systemctl disable --now firewalld &> /dev/null; ${COLOR}"${NAME}} ${VERSION_ID}操作系统，Firewall防火墙已关闭!"${END}; } || ${COLOR}"${NAME}} ${VERSION_ID}操作系统，iptables防火墙已关闭!"${END}
 }
 
 disable_selinux(){
     if [ `getenforce` == "Enforcing" ];then
         sed -ri.bak 's/^(SELINUX=).*/\1disabled/' /etc/selinux/config
         setenforce 0
-        ${COLOR}"${OS_ID} ${OS_RELEASE} SELinux已禁用,请重新启动系统后才能永久生效!"${END}
+        ${COLOR}"${FULL_NAME}操作系统，SELinux已禁用，请重新启动系统后才能永久生效！"${END}
     else
-        ${COLOR}"${OS_ID} ${OS_RELEASE} SELinux已被禁用,不用设置!"${END}
+        ${COLOR}"${FULL_NAME}操作系统，SELinux已被禁用，不用设置！"${END}
     fi
 }
 
 set_swap(){
     if grep -Eqi "noauto" /etc/fstab;then
-        ${COLOR}"${OS_ID} ${OS_RELEASE} swap已被禁用,不用设置!"${END}
+        ${COLOR}"${NAME}} ${VERSION_ID}操作系统，swap已被禁用，不用设置！"${END}
     else
         sed -ri.bak '/swap/s/(.*)(defaults)(.*)/\1\2,noauto\3/g' /etc/fstab
         swapoff -a
-        ${COLOR}"${OS_ID} ${OS_RELEASE} 禁用swap已设置成功,请重启系统后生效!"${END}
+        ${COLOR}"${FULL_NAME}操作系统，禁用swap已设置成功，请重启系统后生效！"${END}
     fi
 }
 
 set_localtime(){
     timedatectl set-timezone Asia/Shanghai
     echo 'Asia/Shanghai' >/etc/timezone
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 系统时区已设置成功,请重启系统后生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，系统时区已设置成功，请重启系统后生效！"${END}
 }
 
 set_limits(){
@@ -322,7 +380,7 @@ root     hard   memlock  32000
 root     soft   msgqueue 8192000
 root     hard   msgqueue 8192000
 EOF
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 优化资源限制参数成功!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，优化资源限制参数成功！"${END}
 }
 
 set_kernel(){
@@ -419,13 +477,13 @@ net.ipv4.tcp_tw_recycle = 0
 EOF
     fi
     sysctl -p &> /dev/null
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 优化内核参数成功!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，优化内核参数成功！"${END}
 }
 
 optimization_ssh(){
     sed -ri.bak -e 's/^#(UseDNS).*/\1 no/' -e 's/^(GSSAPIAuthentication).*/\1 no/' /etc/ssh/sshd_config
     systemctl restart sshd
-    ${COLOR}"${OS_ID} ${OS_RELEASE} SSH已优化完成!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，SSH已优化完成！"${END}
 }
 
 set_ssh_port(){
@@ -434,7 +492,7 @@ set_ssh_port(){
     read -p "请输入端口号: " PORT
     sed -i 's/#Port 22/Port '${PORT}'/' /etc/ssh/sshd_config
 	systemctl restart sshd
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 更改SSH端口号已完成,请重新登陆后生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，更改SSH端口号已完成，请重新登陆后生效！"${END}
 }
 
 set_base_alias(){
@@ -461,7 +519,7 @@ EOF
 alias scandisk="echo '- - -' > /sys/class/scsi_host/host0/scan;echo '- - -' > /sys/class/scsi_host/host1/scan;echo '- - -' > /sys/class/scsi_host/host2/scan"
 EOF
     fi
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 系统别名已设置成功,请重新登陆后生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，系统别名已设置成功，请重新登陆后生效！"${END}
 }
 
 set_alias(){
@@ -502,12 +560,12 @@ func SetTitle()
 endfunc
 autocmd BufNewFile * normal G
 EOF
-    ${COLOR}"${OS_ID} ${OS_RELEASE} vimrc设置完成,请重新系统启动才能生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，vimrc设置完成，请重新系统启动才能生效！"${END}
 }
 
 set_mail(){                                                                                                 
-    rpm -q postfix &> /dev/null || { ${COLOR}"安装postfix服务,请稍等..."${END};yum -y install postfix &> /dev/null; systemctl enable --now postfix &> /dev/null; }
-    rpm -q mailx &> /dev/null || { ${COLOR}"安装mailx服务,请稍等..."${END};yum -y install mailx &> /dev/null; }
+    rpm -q postfix &> /dev/null || { ${COLOR}"安装postfix服务，请稍等......"${END};yum -y install postfix &> /dev/null; systemctl enable --now postfix &> /dev/null; }
+    rpm -q mailx &> /dev/null || { ${COLOR}"安装mailx服务，请稍等......"${END};yum -y install mailx &> /dev/null; }
     read -p "请输入邮箱地址: " MAIL
     read -p "请输入邮箱授权码: " AUTH
     SMTP=`echo ${MAIL} |awk -F"@" '{print $2}'`
@@ -519,7 +577,7 @@ set smtp-auth-password=${AUTH}
 set smtp-auth=login
 set ssl-verify=ignore
 EOF
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 邮件设置完成,请重新登录后才能生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，邮件设置完成，请重新登录后才能生效！"${END}
 }
 
 red(){
@@ -564,7 +622,7 @@ set_ps1_env(){
 }
 
 set_ps1(){
-    TIPS="${COLOR}${OS_ID} ${OS_RELEASE} PS1设置成功,请重新登录生效!${END}"
+    TIPS="${COLOR}${FULL_NAME}操作系统，PS1设置成功，请重新登录生效！${END}"
     while true;do
         echo -e "\E[$[RANDOM%7+31];1m"
         cat <<-EOF
@@ -620,7 +678,7 @@ EOF
             break
             ;;
         *)
-            ${COLOR}"输入错误,请输入正确的数字(1-8)!"${END}
+            ${COLOR}"输入错误，请输入正确的数字(1-8)！"${END}
             ;;
         esac
     done
@@ -637,7 +695,7 @@ set_vim_env(){
     else
         set_vim
     fi
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 默认文本编辑器设置成功,请重新登录生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，默认文本编辑器设置成功，请重新登录生效！"${END}
 }
 
 set_history(){
@@ -651,7 +709,7 @@ set_history_env(){
     else
         set_history
     fi
-    ${COLOR}"${OS_ID} ${OS_RELEASE} history格式设置成功,请重新登录生效!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，history格式设置成功，请重新登录生效！"${END}
 }
 
 disable_restart(){
@@ -660,7 +718,7 @@ disable_restart(){
         systemctl disable ctrl-alt-del.target
     fi
     systemctl mask ctrl-alt-del.target
-    ${COLOR}"${OS_ID} ${OS_RELEASE} 禁用ctrl+alt+del重启功能设置成功!"${END}
+    ${COLOR}"${FULL_NAME}操作系统，禁用ctrl+alt+del重启功能设置成功！"${END}
 }
 
 menu(){
@@ -697,7 +755,7 @@ EOF
             set_hostname
             ;;
         4)
-            base_menu
+            opencloudos_menu
             ;;
         5)
             minimal_install
@@ -757,7 +815,7 @@ EOF
             break
             ;;
         *)
-            ${COLOR}"输入错误,请输入正确的数字(1-23)!"${END}
+            ${COLOR}"输入错误，请输入正确的数字(1-23)！"${END}
             ;;
         esac
     done
@@ -765,10 +823,10 @@ EOF
 
 main(){
     os
-    if [ ${OS_ID} == "OpenCloudOS" ];then
+    if [ ${MAIN_NAME} == "OpenCloudOS" ];then
         menu
     else
-        ${COLOR}"此脚本不支持${OS_ID} ${OS_RELEASE} 系统!"${END}
+        ${COLOR}"此脚本不支持${FULL_NAME}操作系统！"${END}
     fi
 }
